Changed notch_skill( char, skill, chance ) such that chance no longer represents a 1 in chance+1 chance to notch, but now is a % chance to notch.  Also, chance is now a float so we can send a 15.25/2.5/etc % chance if we want to.

+++ src/prototypes.h
@@ -1078,7 +1078,7 @@ void do_remort(P_char, char *, int);
 void do_spec(P_char, char *, int);
 void do_skills(P_char, char *, int);
 void do_spells(P_char, char *, int);
-int notch_skill(P_char, int, int);
+int notch_skill(P_char, int, float);
 void SetGuildSpellLvl(void);
 void update_skills(P_char);
